requirements.txt
# Please keep these requirements alphabetized

flask
pathlib
python-dotenv
requests